+++ packages/eslint-config-nikita/rules/variables.js
@@ -10,11 +10,11 @@ module.exports = {
         'no-delete-var': 'error',
 
         // disallow labels that share a name with a variable
-        // http://eslint.org/docs/rules/no-label-var
+        // https://eslint.org/docs/rules/no-label-var
         'no-label-var': 'error',
 
         // disallow specific globals
-        'no-restricted-globals': 'off',
+        'no-restricted-globals': ['error', 'isFinite', 'isNaN'],
 
         // disallow declaration of variables already declared in the outer scope
         'no-shadow': 'error',
@@ -29,11 +29,11 @@ module.exports = {
         'no-undef-init': 'error',
 
         // disallow use of undefined variable
-        // http://eslint.org/docs/rules/no-undefined
+        // https://eslint.org/docs/rules/no-undefined
         'no-undefined': 'error',
 
         // disallow declaration of variables that are not used in the code
-        'no-unused-vars': ['warn', { vars: 'local', args: 'after-used', ignoreRestSiblings: true }],
+        'no-unused-vars': ['warn', { vars: 'all', args: 'after-used', ignoreRestSiblings: true }],
 
         // disallow use of variables before they are defined
         'no-use-before-define': ['error', { functions: true, classes: true, variables: true }],